updated user functions
